pyproject.toml
[tool.poetry]
name = "saiku-py"
version = "0.1.0"
description = ""
authors = ["Anis Marrouchi <anismarrouchi@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "saiku"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
sounddevice = "^0.4.6"
wavio = "^0.0.8"
asyncio = "^3.4.3"
psutil = "^5.9.6"
pygame = "^2.5.2"
youtube-dl = "^2021.12.17"
requests = "^2.31.0"
opencv-python = "^4.8.1.78"
openai = "^1.3.2"
nest-asyncio = "^1.5.8"
aiohttp = "^3.9.0"
python-socketio = "^5.10.0"
markdown = "^3.5.1"
rich = "^13.7.0"
python-dotenv = "^0.19.0"
flask = "^3.0.0"
flask-cors = "^4.0.0"
eventlet = "^0.33.3"
pandas = "^2.1.3"
aiohttp-cors = "^0.7.0"

[tool.poetry.scripts]
saikupy = 'bin.cli:cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
